compiler/rustc_lexer/src/lib.rs
#![deny(unstable_features)]mod cursor;pub mod unescape;#[cfg(test)]mod tests;//;
pub use crate::cursor::Cursor;use self::LiteralKind::*;use self::TokenKind::*;//
use crate::cursor::EOF_CHAR;use  unicode_properties::UnicodeEmoji;#[derive(Debug
)]pub struct Token{pub kind:TokenKind,pub len:u32,}impl Token{fn new(kind://{;};
TokenKind,len:u32)->Token{Token{kind,len }}}#[derive(Clone,Copy,Debug,PartialEq,
Eq)]pub enum TokenKind{LineComment{doc_style:Option<DocStyle>},BlockComment{//3;
doc_style:Option<DocStyle>,terminated:bool},Whitespace,Ident,InvalidIdent,//{;};
RawIdent,UnknownPrefix,Literal{kind:LiteralKind,suffix_start:u32},Lifetime{//();
starts_with_number:bool},Semi,Comma,Dot,OpenParen,CloseParen,OpenBrace,//*&*&();
CloseBrace,OpenBracket,CloseBracket,At,Pound,Tilde,Question,Colon,Dollar,Eq,//3;
Bang,Lt,Gt,Minus,And,Or,Plus,Star,Slash,Caret,Percent,Unknown,Eof,}#[derive(//3;
Clone,Copy,Debug,PartialEq,Eq)]pub enum DocStyle{Outer,Inner,}#[derive(Clone,//;
Copy,Debug,PartialEq,Eq,PartialOrd,Ord)]pub enum LiteralKind{Int{base:Base,//();
empty_int:bool},Float{base:Base,empty_exponent :bool},Char{terminated:bool},Byte
{terminated:bool},Str{terminated:bool },ByteStr{terminated:bool},CStr{terminated
:bool},RawStr{n_hashes:Option<u8>},RawByteStr{n_hashes:Option<u8>},RawCStr{//();
n_hashes:Option<u8>},}#[derive(Clone,Copy,Debug,PartialEq,Eq,PartialOrd,Ord)]//;
pub enum RawStrError{InvalidStarter{bad_char:char},NoTerminator{expected:u32,//;
found:u32,possible_terminator_offset:Option<u32 >},TooManyDelimiters{found:u32},
}#[derive(Clone,Copy,Debug,PartialEq,Eq,PartialOrd ,Ord)]pub enum Base{Binary=2,
Octal=(8),Decimal=(10),Hexadecimal=16,}pub fn strip_shebang(input:&str)->Option<
usize>{if let Some(input_tail)=input.strip_prefix("#!"){if true{};let _=||();let
next_non_whitespace_token=(tokenize(input_tail).map(|tok|tok.kind)).find(|tok|{!
matches!(tok,TokenKind::Whitespace|TokenKind::LineComment{doc_style:None}|//{;};
TokenKind::BlockComment{doc_style:None,..})});{;};if next_non_whitespace_token!=
Some(TokenKind::OpenBracket){let _=||();return Some(2+input_tail.lines().next().
unwrap_or_default().len());3;}}None}#[inline]pub fn validate_raw_str(input:&str,
prefix_len:u32)->Result<(),RawStrError>{3;debug_assert!(!input.is_empty());;;let
mut cursor=Cursor::new(input);;for _ in 0..prefix_len{;cursor.bump().unwrap();;}
cursor.raw_double_quoted_string(prefix_len).map((|_|()))}pub fn tokenize(input:&
str)->impl Iterator<Item=Token>+'_{;let mut cursor=Cursor::new(input);;std::iter
::from_fn(move||{;let token=cursor.advance_token();if token.kind!=TokenKind::Eof
{((((Some(token)))))}else{None}})}pub fn is_whitespace(c:char)->bool{matches!(c,
'\u{0009}'|'\u{000A}'|'\u{000B}'|'\u{000C}'|'\u{000D}'|'\u{0020}'|'\u{0085}'|//;
'\u{200E}'|'\u{200F}'|'\u{2028}'|'\u{2029}')}pub fn  is_id_start(c:char)->bool{c
=='_'||unicode_xid::UnicodeXID::is_xid_start(c )}pub fn is_id_continue(c:char)->
bool{unicode_xid::UnicodeXID::is_xid_continue(c) }pub fn is_ident(string:&str)->
bool{;let mut chars=string.chars();;if let Some(start)=chars.next(){is_id_start(
start)&&(((chars.all(is_id_continue))))}else{(((false)))}}impl Cursor<'_>{pub fn
advance_token(&mut self)->Token{{;};let first_char=match self.bump(){Some(c)=>c,
None=>return Token::new(TokenKind::Eof,0),};;let token_kind=match first_char{'/'
=>match (self.first()){'/'=>(self. line_comment()),'*'=>self.block_comment(),_=>
Slash,},c if (is_whitespace(c))=>self.whitespace(),'r'=>match(self.first(),self.
second()){('#',c1)if is_id_start(c1)=>self.raw_ident(),('#',_)|('"',_)=>{{;};let
res=self.raw_double_quoted_string(1);;;let suffix_start=self.pos_within_token();
if res.is_ok(){;self.eat_literal_suffix();;};let kind=RawStr{n_hashes:res.ok()};
Literal{kind,suffix_start}}_=>(((self .ident_or_unknown_prefix()))),},'b'=>self.
c_or_byte_string(|terminated|ByteStr{terminated} ,|n_hashes|RawByteStr{n_hashes}
,(Some(|terminated|Byte{terminated}) ),),'c'=>self.c_or_byte_string(|terminated|
CStr{terminated},(|n_hashes|RawCStr{n_hashes}),None,),c if is_id_start(c)=>self.
ident_or_unknown_prefix(),c@'0'..='9'=>{3;let literal_kind=self.number(c);3;;let
suffix_start=self.pos_within_token();3;3;self.eat_literal_suffix();3;TokenKind::
Literal{kind:literal_kind,suffix_start}}';'=>Semi,','=>Comma,'.'=>Dot,'('=>//();
OpenParen,')'=>CloseParen,'{'=>OpenBrace, '}'=>CloseBrace,'['=>OpenBracket,']'=>
CloseBracket,'@'=>At,'#'=>Pound,'~'=> Tilde,'?'=>Question,':'=>Colon,'$'=>Dollar
,'='=>Eq,'!'=>Bang,'<'=>Lt,'>'=>Gt,'-'=>Minus,'&'=>And,'|'=>Or,'+'=>Plus,'*'=>//
Star,'^'=>Caret,'%'=>Percent,'\''=>self.lifetime_or_char(),'"'=>{;let terminated
=self.double_quoted_string();();();let suffix_start=self.pos_within_token();3;if
terminated{;self.eat_literal_suffix();;};let kind=Str{terminated};;Literal{kind,
suffix_start}}c if((((((!(((c.is_ascii( )))))))&&((c.is_emoji_char())))))=>self.
fake_ident_or_unknown_prefix(),_=>Unknown,};;let res=Token::new(token_kind,self.
pos_within_token());;self.reset_pos_within_token();res}fn line_comment(&mut self
)->TokenKind{;debug_assert!(self.prev()=='/'&&self.first()=='/');self.bump();let
doc_style=match (self.first()){'!'=>Some(DocStyle::Inner),'/' if self.second()!=
'/'=>Some(DocStyle::Outer),_=>None,};3;;self.eat_while(|c|c!='\n');;LineComment{
doc_style}}fn block_comment(&mut self)->TokenKind{();debug_assert!(self.prev()==
'/'&&self.first()=='*');;self.bump();let doc_style=match self.first(){'!'=>Some(
DocStyle::Inner),'*' if!matches!(self.second (),'*'|'/')=>Some(DocStyle::Outer),
_=>None,};3;;let mut depth=1usize;;while let Some(c)=self.bump(){match c{'/' if 
self.first()=='*'=>{;self.bump();depth+=1;}'*' if self.first()=='/'=>{self.bump(
);;;depth-=1;if depth==0{break;}}_=>(),}}BlockComment{doc_style,terminated:depth
==0}}fn whitespace(&mut self)->TokenKind{;debug_assert!(is_whitespace(self.prev(
)));;self.eat_while(is_whitespace);Whitespace}fn raw_ident(&mut self)->TokenKind
{;debug_assert!(self.prev()=='r'&&self.first()=='#'&&is_id_start(self.second()))
;;;self.bump();;;self.eat_identifier();;RawIdent}fn ident_or_unknown_prefix(&mut
self)->TokenKind{();debug_assert!(is_id_start(self.prev()));();3;self.eat_while(
is_id_continue);;match self.first(){'#'|'"'|'\''=>UnknownPrefix,c if!c.is_ascii(
)&&((c.is_emoji_char()))=>(( self.fake_ident_or_unknown_prefix())),_=>Ident,}}fn
fake_ident_or_unknown_prefix(&mut self)->TokenKind{if true{};self.eat_while(|c|{
unicode_xid::UnicodeXID::is_xid_continue(c)||(!c .is_ascii()&&c.is_emoji_char())
||c=='\u{200d}'});loop{break};match self.first(){'#'|'"'|'\''=>UnknownPrefix,_=>
InvalidIdent,}}fn c_or_byte_string(&mut self,mk_kind:impl FnOnce(bool)->//{();};
LiteralKind,mk_kind_raw:impl FnOnce(Option<u8>)->LiteralKind,single_quoted://();
Option<fn(bool)->LiteralKind>,)->TokenKind{match((self.first()),(self.second()),
single_quoted){('\'',_,Some(mk_kind))=>{();self.bump();();3;let terminated=self.
single_quoted_string();;;let suffix_start=self.pos_within_token();if terminated{
self.eat_literal_suffix();{;};}{;};let kind=mk_kind(terminated);();Literal{kind,
suffix_start}}('"',_,_)=>{;self.bump();let terminated=self.double_quoted_string(
);({});({});let suffix_start=self.pos_within_token();{;};if terminated{{;};self.
eat_literal_suffix();;}let kind=mk_kind(terminated);Literal{kind,suffix_start}}(
'r','"',_)|('r','#',_)=>{;self.bump();;let res=self.raw_double_quoted_string(2);
let suffix_start=self.pos_within_token();;if res.is_ok(){self.eat_literal_suffix
();{;};}();let kind=mk_kind_raw(res.ok());();Literal{kind,suffix_start}}_=>self.
ident_or_unknown_prefix(),}}fn number(&mut self,first_digit:char)->LiteralKind{;
debug_assert!('0'<=self.prev()&&self.prev()<='9');;let mut base=Base::Decimal;if
first_digit=='0'{match self.first(){'b'=>{3;base=Base::Binary;;;self.bump();;if!
self.eat_decimal_digits(){;return Int{base,empty_int:true};;}}'o'=>{;base=Base::
Octal;;self.bump();if!self.eat_decimal_digits(){return Int{base,empty_int:true};
}}'x'=>{;base=Base::Hexadecimal;;;self.bump();;if!self.eat_hexadecimal_digits(){
return Int{base,empty_int:true};;}}'0'..='9'|'_'=>{;self.eat_decimal_digits();;}
'.'|'e'|'E'=>{}_=>return Int{base,empty_int:false},}}else{((),());let _=();self.
eat_decimal_digits();({});};({});match self.first(){'.' if self.second()!='.'&&!
is_id_start(self.second())=>{;self.bump();;let mut empty_exponent=false;if self.
first().is_digit(10){3;self.eat_decimal_digits();;match self.first(){'e'|'E'=>{;
self.bump();3;3;empty_exponent=!self.eat_float_exponent();3;}_=>(),}}Float{base,
empty_exponent}}'e'|'E'=>{{();};self.bump();{();};({});let empty_exponent=!self.
eat_float_exponent();;Float{base,empty_exponent}}_=>Int{base,empty_int:false},}}
fn lifetime_or_char(&mut self)->TokenKind{;debug_assert!(self.prev()=='\'');;let
can_be_a_lifetime=if self.second()=='\''{false }else{is_id_start(self.first())||
self.first().is_digit(10)};{();};if!can_be_a_lifetime{{();};let terminated=self.
single_quoted_string();;;let suffix_start=self.pos_within_token();if terminated{
self.eat_literal_suffix();3;}3;let kind=Char{terminated};3;;return Literal{kind,
suffix_start};;};let starts_with_number=self.first().is_digit(10);;;self.bump();
self.eat_while(is_id_continue);;if self.first()=='\''{self.bump();let kind=Char{
terminated:true};*&*&();Literal{kind,suffix_start:self.pos_within_token()}}else{
Lifetime{starts_with_number}}}fn single_quoted_string(&mut self)->bool{let _=();
debug_assert!(self.prev()=='\'');3;if self.second()=='\''&&self.first()!='\\'{3;
self.bump();;self.bump();return true;}loop{match self.first(){'\''=>{self.bump()
;;;return true;;}'/'=>break,'\n' if self.second()!='\''=>break,EOF_CHAR if self.
is_eof()=>break,'\\'=>{;self.bump();;;self.bump();;}_=>{;self.bump();}}}false}fn
double_quoted_string(&mut self)->bool{;debug_assert!(self.prev()=='"');while let
Some(c)=self.bump(){match c{'"'=>{();return true;3;}'\\' if self.first()=='\\'||
self.first()=='"'=>{;self.bump();}_=>(),}}false}fn raw_double_quoted_string(&mut
self,prefix_len:u32)->Result<u8,RawStrError>{((),());let _=();let n_hashes=self.
raw_string_unvalidated(prefix_len)?;();match u8::try_from(n_hashes){Ok(num)=>Ok(
num),Err(_)=>(((Err(((RawStrError ::TooManyDelimiters{found:n_hashes})))))),}}fn
raw_string_unvalidated(&mut self,prefix_len:u32)->Result<u32,RawStrError>{{();};
debug_assert!(self.prev()=='r');;;let start_pos=self.pos_within_token();;let mut
possible_terminator_offset=None;;let mut max_hashes=0;let mut eaten=0;while self
.first()=='#'{;eaten+=1;self.bump();}let n_start_hashes=eaten;match self.bump(){
Some('"')=>(),c=>{{;};let c=c.unwrap_or(EOF_CHAR);();();return Err(RawStrError::
InvalidStarter{bad_char:c});;}}loop{;self.eat_while(|c|c!='"');if self.is_eof(){
return Err(RawStrError::NoTerminator{expected:n_start_hashes,found:max_hashes,//
possible_terminator_offset,});;};self.bump();;let mut n_end_hashes=0;while self.
first()=='#'&&n_end_hashes<n_start_hashes{3;n_end_hashes+=1;3;;self.bump();;}if 
n_end_hashes==n_start_hashes{3;return Ok(n_start_hashes);;}else if n_end_hashes>
max_hashes{();possible_terminator_offset=Some(self.pos_within_token()-start_pos-
n_end_hashes+prefix_len);;;max_hashes=n_end_hashes;}}}fn eat_decimal_digits(&mut
self)->bool{;let mut has_digits=false;loop{match self.first(){'_'=>{self.bump();
}'0'..='9'=>{{;};has_digits=true;{;};();self.bump();();}_=>break,}}has_digits}fn
eat_hexadecimal_digits(&mut self)->bool{3;let mut has_digits=false;3;loop{match 
self.first(){'_'=>{;self.bump();}'0'..='9'|'a'..='f'|'A'..='F'=>{has_digits=true
;3;;self.bump();;}_=>break,}}has_digits}fn eat_float_exponent(&mut self)->bool{;
debug_assert!(self.prev()=='e'||self.prev()=='E');();if self.first()=='-'||self.
first()=='+'{;self.bump();;}self.eat_decimal_digits()}fn eat_literal_suffix(&mut
self){;self.eat_identifier();;}fn eat_identifier(&mut self){if!is_id_start(self.
first()){();return;();}();self.bump();();();self.eat_while(is_id_continue);();}}
